[Update] Input handling
Updates input handling to be handled
by the Scene instead.
For now, just the mouse input is being handled,
and the keyboard input for the player is still
handled by the player itself.
This will be changed later.
Updated buttons (now the hovered property
is private, and handled by methods instead)