[Remove] GameObject
As wrote each new class I noticed it would stray
too far from the original GameObject, and so,
I ended up implementing basically all the classes
without GameObject, as it was more of a limitation
than something helpful.
Maybe I will bring it back in the future with a
new formula. As it stood, it was just useless.

Button class updated to remove it's (minimal)
dependency on GameObject.

Scene class updated to reflect the changes.