[Update] Collectables as a list
Now collectables are a list, meaning a single
quest can have multiple collectables.
Now collectables have much fewer properties, as
most value calculations are done in the JSON
parsing instead, to get rid of the need to pass the
level's tile size and scale to the collectables.
Slight cleanup on level and scene updates.